refactor: separate tests
