+++ tests/basic-spec.js
@@ -1,0 +1,7 @@
+export default (angular2now, ngModuleName) => {
+  describe("basic", () => {
+    it("should be available defined", () => {
+      expect(angular2now).toBeDefined();
+    });
+  });
+};
