+++ tests/component-spec.js
@@ -1,0 +1,458 @@
+export default (angular2now, ngModuleName) => {
+  describe("@Component()", () => {
+    const moduleMock = {
+      directive: () => {}
+    };
+    let target;
+    let spyModule;
+    let spyDirective;
+    const nameCamel = 'testComponent';
+    const nameClass = '.testComponent';
+    const nameDashed = 'test-component';
+
+    /**
+     * Returns directive's object
+     * @param  {Object|String} opt Component(opt)
+     * @return {Object} target reference
+     */
+    function getDDO(opt) {
+      const ddo = spyDirective.calls.mostRecent().args[1]();
+
+      return ddo[opt];
+    }
+
+    /**
+     * Retuns directive's name
+     * @return {[type]} [description]
+     */
+    function getDDName() {
+      return spyDirective.calls.mostRecent().args[0];
+    }
+
+    /**
+     * Shorthand for angular2now.Component(opts)(target)
+     * @param  {Object|String} opts Component(opts)
+     * @return {Object} target reference
+     */
+    function doComponent(opts) {
+      return angular2now.Component(opts)(target);
+    }
+
+    beforeEach(() => {
+      target = function target() {};
+      spyModule = spyOn(angular, 'module').and.returnValue(moduleMock);
+      spyDirective = spyOn(moduleMock, 'directive');
+    });
+
+    it("should have target at controller", () => {
+      doComponent(nameDashed);
+
+      expect(getDDO('controller')).toBe(target);
+    });
+
+    describe("options.selector", () => {
+
+      it("should set selector if argument is a string", () => {
+        doComponent(nameDashed);
+
+        expect(target.selector).toBe(nameDashed);
+        expect(getDDName()).toBe(nameCamel);
+      });
+
+      it("should be able to unCamelCase selector", () => {
+        doComponent(nameCamel);
+
+        expect(target.selector).toBe(nameDashed);
+        expect(getDDName()).toBe(nameCamel);
+      });
+
+      it("should handle class name as selector", () => {
+        doComponent(nameClass);
+
+        expect(target.selector).toBe(nameDashed);
+        expect(getDDName()).toBe(nameCamel);
+      });
+
+      it("should call angular.directive with proper selector", () => {
+        doComponent(nameClass);
+
+        expect(target.selector).toBe(nameDashed);
+        expect(spyDirective).toHaveBeenCalled();
+        expect(getDDName()).toBe(nameCamel);
+      });
+    });
+
+    describe("options.injectables", () => {
+      const injectables = ['$http', '$q'];
+
+      it("should set injectables", () => {
+        doComponent({
+          injectables
+        });
+
+        expect(target.$inject).toEqual(injectables);
+        expect(target.$injectDefer).toEqual(target.$inject);
+      });
+    });
+
+    describe("options.services", () => {
+      const services = ['$http', '$q'];
+
+      it("should set services", () => {
+        doComponent({
+          services
+        });
+
+        expect(target.$inject).toEqual(services);
+        expect(target.$injectDefer).toEqual(target.$inject);
+      });
+    });
+
+    describe("options.template", () => {
+      it("should set template", () => {
+        const template = 'foo';
+
+        doComponent({
+          template
+        });
+
+        expect(getDDO('template')).toBe(template);
+      });
+    });
+
+    describe("options.templateUrl", () => {
+      it("should set templateUrl", () => {
+        const templateUrl = 'foo.html';
+
+        doComponent({
+          templateUrl
+        });
+
+        expect(getDDO('templateUrl')).toBe(templateUrl);
+      });
+    });
+
+    describe("options.transclude", () => {
+      it("should set transclude to true", () => {
+        const transclude = true;
+
+        doComponent({
+          transclude
+        });
+
+        expect(getDDO('transclude')).toBe(transclude);
+      });
+      it("should set not set transclude", () => {
+        const transclude = false;
+
+        doComponent({
+          transclude
+        });
+
+        expect(getDDO('transclude')).toBeUndefined();
+      });
+    });
+
+    describe("options.restrict", () => {
+
+      it("should set each restriction", () => {
+        const restrictions = ['E', 'A', 'C', 'EA', 'EAC'];
+
+        restrictions.forEach((restrict) => {
+          doComponent({
+            restrict
+          });
+
+          expect(getDDO('restrict')).toBe(restrict);
+        });
+      });
+    });
+
+    describe("options.controllerAs", () => {
+      it("should set controllerAs", () => {
+        const controllerAs = 'foo';
+
+        doComponent({
+          controllerAs
+        });
+
+        expect(getDDO('controllerAs')).toBe(controllerAs);
+      });
+
+      it("should overwrite target's controllerAs", () => {
+        const controllerAs = 'foo';
+
+        target.controllerAs = 'bar';
+        doComponent({
+          controllerAs
+        });
+
+        expect(getDDO('controllerAs')).toBe(controllerAs);
+      });
+    });
+
+    describe("options.scope", () => {
+      it("should set scope", () => {
+        const scopes = [true, false, undefined, {
+          foo: 'bar'
+        }];
+
+        scopes.forEach((scope) => {
+          doComponent({
+            scope
+          });
+
+          expect(getDDO('scope')).toBe(scope);
+        });
+      });
+
+      it("should be skipped if target's scope is available", () => {
+        const scopes = [true, false, undefined, {
+          foo: 'bar'
+        }];
+
+        scopes.forEach((scope) => {
+          target.scope = scope;
+          doComponent({
+            scope: {
+              foo: 'baz'
+            }
+          });
+
+          expect(getDDO('scope')).toBe(scope);
+        });
+      });
+
+      it("should be an empty object if neither of bind, scope or target's scope is available", () => {
+        doComponent({
+          selector: nameDashed
+        });
+
+        expect(getDDO('scope')).toEqual({});
+      });
+    });
+
+    describe("target.bindToController", () => {
+      it("shoud set bindToController", () => {
+        const bools = [true, false];
+
+        bools.forEach((val) => {
+          target.bindToController = val;
+          doComponent();
+
+          expect(getDDO('bindToController')).toBe(val);
+        });
+      });
+    });
+
+    describe("target.require", () => {
+      it("should set require", () => {
+        const require = ['@foo'];
+
+        target.require = require;
+        doComponent();
+
+        expect(getDDO('require')).toBe(require);
+      });
+    });
+
+    describe("options.require", () => {
+      it("should set require", () => {
+        const require = ['@foo'];
+
+        doComponent({
+          require
+        });
+
+        expect(getDDO('require')).toBe(require);
+      });
+
+      it("should overwrite target's require", () => {
+        const require = ['@foo'];
+
+        target.require = ['@bar'];
+        doComponent({
+          require
+        });
+
+        expect(getDDO('require')).toBe(require);
+      });
+    });
+
+    describe("require", () => {
+      const monkeyInjects = ['@^foo', '@bar'];
+      const injectables = ['$http'].concat(monkeyInjects);
+      const questionInjects = monkeyInjects.map((inj) => '?' + inj.slice(1));
+
+      it("should set require with @ prefixed injectables", () => {
+        doComponent({
+          injectables,
+          selector: nameDashed
+        });
+
+        const requireDDO = getDDO('require');
+
+        questionInjects.forEach((inj) => {
+          expect(requireDDO.indexOf(inj)).not.toBe(-1);
+        });
+      });
+
+      it("should set require with target", () => {
+        doComponent({
+          injectables,
+          selector: nameDashed
+        });
+
+        expect(getDDO('require').indexOf(nameCamel)).not.toBe(-1);
+      });
+
+      it("should remove transformed injectables", () => {
+        doComponent({
+          injectables,
+          selector: nameDashed
+        });
+
+        monkeyInjects.forEach((inj) => {
+          expect(target.$inject.indexOf(inj)).toBe(-1);
+          expect(target.$injectDefer.indexOf(inj)).toBe(-1);
+        });
+      });
+    });
+
+    describe("target.link", () => {
+      it("should set link", () => {
+        const link = function linkTarget() {};
+
+        target.link = link;
+        doComponent();
+
+        expect(getDDO('link')).toBe(link);
+      });
+    });
+
+    describe("options.link", () => {
+      it("should set link", () => {
+        const link = function linkOptions() {};
+
+        doComponent({
+          link
+        });
+
+        expect(getDDO('link')).toBe(link);
+      });
+
+      it("should overwrite target's link", () => {
+        const link = function linkOptions() {};
+
+        target.link = function linkTarget() {};
+        doComponent({
+          link
+        });
+
+        expect(getDDO('link')).toBe(link);
+      });
+    });
+
+    describe("restrict", () => {
+      it("should be A by default", () => {
+        doComponent(nameCamel);
+        expect(getDDO('restrict')).toBe('A');
+      });
+
+      it("should be EA when templateUrl and non class selector", () => {
+        doComponent({
+          selector: nameCamel,
+          templateUrl: 'foo.html'
+        });
+        expect(getDDO('restrict')).toBe('EA');
+      });
+
+      it("should be EA when template and non class selector", () => {
+        doComponent({
+          selector: nameCamel,
+          template: 'foo'
+        });
+        expect(getDDO('restrict')).toBe('EA');
+      });
+
+      it("should be C when slector is a class name", () => {
+        doComponent({
+          selector: nameClass
+        });
+        expect(getDDO('restrict')).toBe('C');
+      });
+    });
+
+    describe("controllerAs", () => {
+      it("should be the same as selector by default", () => {
+        doComponent({
+          selector: nameDashed
+        });
+        expect(getDDO('controllerAs')).toBe(nameCamel);
+      });
+
+      it("should be owerwritten by options.controllerAs", () => {
+        angular2now.options({
+          controllerAs: 'vm'
+        });
+        doComponent({
+          selector: nameDashed
+        });
+        expect(getDDO('controllerAs')).toBe('vm');
+
+        // reset
+        angular2now.options({
+          controllerAs: null
+        });
+      });
+    });
+
+    describe("options.bind", () => {
+      it("should set scope", () => {
+        const bind = {
+          foo: 'bar'
+        };
+
+        doComponent({
+          bind
+        });
+
+        expect(getDDO('scope')).toEqual(bind);
+      });
+    });
+
+    describe("transclude", () => {
+      it("should be undefined by default", () => {
+        doComponent({
+          selector: nameDashed
+        });
+
+        expect(getDDO('transclude')).toBeUndefined();
+      });
+
+      it("should set transclude when template contains content element", () => {
+        const template = `
+        <div>
+          <content></content>
+        </div>
+      `;
+
+        doComponent({
+          template
+        });
+
+        expect(getDDO('transclude')).toBe(true);
+      });
+
+      it("should set transclude when template does not contain content", () => {
+        const template = `<div></div>`;
+
+        doComponent({
+          template
+        });
+
+        expect(getDDO('transclude')).toBeUndefined();
+      });
+    });
+  });
+};
